Fix right click near top edge of maximized windows
Previously, right clicking the very top of a maximized window wouldn't
produce the drop down menu because this area was reserved for resizing.
This takes into account the fact that the window is maximized and
ignores any upper padding, since maximized windows shouldn't be able to
be resized by the user anyway.

This doesn't handle the literal corner cases, that will have to be for
a later patch.